+++ ricaun.NUnit/Revit/Commands/Command.cs
@@ -1,54 +1,86 @@
-#if DEBUG
 using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
+using Microsoft.Win32;
 using System;
+using System.Diagnostics;
 using System.IO;
-using System.IO.Compression;
 using System.Linq;
 using System.Reflection;
+using System.Windows.Input;
 
 namespace ricaun.NUnit.Revit.Commands
 {
     [Transaction(TransactionMode.Manual)]
     public class Command : IExternalCommand
     {
+        public static string LastPath { get; set; } = @"D:\Users\ricau\Documents\GitHub\ricaun-io\ricaun.NUnit\SampleTest.Tests\bin\Debug\SampleTest.Tests.dll";
         public Result Execute(ExternalCommandData commandData, ref string message, ElementSet elementSet)
         {
             UIApplication uiapp = commandData.Application;
 
-            var location = Assembly.GetExecutingAssembly().Location;
-            var directory = Path.GetDirectoryName(location);
+            if (!File.Exists(LastPath)) LastPath = null;
 
-            //foreach (var item in Directory.GetFiles(Path.GetDirectoryName(location), "*.dll"))
-            //{
-            //    Console.WriteLine($"{TestEngine.ContainNUnit(item)} {Path.GetFileName(item)}");
-            //}
-            //var test = TestEngine.TestAssembly(location, "This is a custom string parameter.", uiapp);
-            //var json = Newtonsoft.Json.JsonConvert.SerializeObject(test);
-            //System.Windows.Clipboard.SetText(json);
-            //Console.WriteLine(test);
+            if (LastPath is null)
+            {
+                OpenFileDialog openFileDialog = new OpenFileDialog();
+                openFileDialog.Filter = "All files (*.*)|*.*";
+                openFileDialog.RestoreDirectory = true;
 
+                if (openFileDialog.ShowDialog() == true)
+                {
+                    //Get the path of specified file
+                    LastPath = openFileDialog.FileName;
+                }
+            }
 
-            UnZipAndTestFiles(uiapp.Application, directory);
-
-
-
+            Execute(uiapp, LastPath);
 
 
             return Result.Succeeded;
         }
 
-
-        private static void CopyFile(string filePath, string directory)
+        private static void Execute(UIApplication uiapp, string filePath)
         {
-            File.Copy(filePath, Path.Combine(directory, Path.GetFileName(filePath)), true);
+            if (filePath is null) return;
+
+            var location = Assembly.GetExecutingAssembly().Location;
+            var directory = Path.GetDirectoryName(location);
+
+            string copyPath = null;
+            if (Path.GetExtension(filePath).EndsWith("zip"))
+            {
+                copyPath = CopyFile(filePath, directory);
+            }
+            else if (Path.GetExtension(filePath).EndsWith("dll"))
+            {
+                copyPath = ZipExtension.CreateFromDirectory(
+                    Path.GetDirectoryName(filePath),
+                    Path.Combine(directory, Path.GetFileName(filePath))
+                    );
+            }
+            else return;
+
+            if (System.Windows.Input.Keyboard.Modifiers == ModifierKeys.Control)
+                Process.Start(directory);
+
+            UnZipAndTestFiles(uiapp, directory);
+
+            if (copyPath is not null)
+                File.Delete(copyPath);
         }
 
-        private static void UnZipAndTestFiles(Application application, string directory)
+        private static string CopyFile(string filePath, string directory)
         {
-            var versionNumber = application.VersionNumber;
+            var copy = Path.Combine(directory, Path.GetFileName(filePath));
+            File.Copy(filePath, copy, true);
+            return copy;
+        }
+
+        private static void UnZipAndTestFiles(UIApplication uiapp, string directory)
+        {
+            var versionNumber = uiapp.Application.VersionNumber;
 
             if (Directory.GetFiles(directory, "*.zip").FirstOrDefault() is string zipFile)
             {
@@ -59,33 +91,36 @@ namespace ricaun.NUnit.Revit.Commands
                         if (Path.GetFileName(versionDirectory).Equals(versionNumber))
                         {
                             Console.WriteLine($"Test VersionNumber: {versionNumber}");
-                            TestDirectory(application, versionDirectory);
+                            TestDirectory(uiapp, versionDirectory);
                         }
                     }
 
-                    TestDirectory(application, zipDestination);
+                    TestDirectory(uiapp, zipDestination);
                 }
             }
         }
 
-        private static void TestDirectory(Application application, string directory)
+        private static void TestDirectory(UIApplication uiapp, string directory)
         {
+            var application = uiapp.Application;
             foreach (var filePath in Directory.GetFiles(directory, "*.dll"))
             {
                 var fileName = Path.GetFileName(filePath);
-                Console.WriteLine($"Test File: {fileName}");
                 try
                 {
                     if (TestEngine.ContainNUnit(filePath))
                     {
+                        Console.WriteLine($"Test File: {fileName}");
+
                         var modelTest = TestEngine.TestAssembly(
                             filePath,
                             application,
-                            application.GetControlledApplication());
+                            application.GetControlledApplication(),
+                            uiapp);
 
                         System.Windows.Clipboard.SetText(Newtonsoft.Json.JsonConvert.SerializeObject(modelTest));
 
-                        Console.WriteLine(modelTest);
+                        Console.WriteLine($"\t{modelTest}");
                     }
 
                 }
@@ -96,59 +131,4 @@ namespace ricaun.NUnit.Revit.Commands
             }
         }
     }
-
-    public static class ZipExtension
-    {
-        /// <summary>
-        /// Get <see cref="ControlledApplication"/> using the <paramref name="application"/>
-        /// </summary>
-        /// <param name="application"></param>
-        /// <returns></returns>
-        public static ControlledApplication GetControlledApplication(this Application application)
-        {
-            var type = typeof(ControlledApplication);
-
-            var constructor = type.GetConstructor(
-                BindingFlags.Instance | BindingFlags.NonPublic, null, new Type[] { application.GetType() }, null);
-
-            return constructor?.Invoke(new object[] { application }) as ControlledApplication;
-        }
-
-        /// <summary>
-        /// Extract <paramref name="zipFile"/> to Folder
-        /// </summary>
-        /// <param name="zipFile"></param>
-        /// <param name="zipDestination"></param>
-        /// <returns></returns>
-        public static bool ExtractToFolder(string zipFile, out string zipDestination)
-        {
-            var zipName = Path.GetFileNameWithoutExtension(zipFile);
-            var zipDirectory = Path.GetDirectoryName(zipFile);
-            zipDestination = Path.Combine(zipDirectory, zipName);
-
-            if (Directory.Exists(zipDestination))
-            {
-                try
-                {
-                    Directory.Delete(zipDestination, true);
-                }
-                catch { }
-            }
-
-            if (Directory.Exists(zipDestination))
-                return true;
-
-            Directory.CreateDirectory(zipDestination);
-
-            try
-            {
-                ZipFile.ExtractToDirectory(zipFile, zipDestination);
-                return true;
-            }
-            catch { }
-
-            return false;
-        }
-    }
 }
-#endif